Refactored the Scumbler::Connect() member fn.
Much of the logic between the connect/disconnect cases was going
to be identical, so rather than create opportunities for copy/paste
errors, I created an internal HandleConnection() method that makes
use of a pointer to member fn to actuall call either
AudioProcessorGraph::addConnection or removeConnection. Other logic
can be inserted inline as appropriate.