Slot drawing partially fixed.
I had forgotten that when you're painting a component, you can draw assuming
that (0, 0) is the upper left of the current component, not the uppoer left
of the parent component. That's what I was doing wrong before (and why only
the first slot was appearing at all, because there happened to be an
intersection between where we were trying to draw and the size of that first
slot component.